Worldinput/Worldedit serialization fixes
Added equality comparisons to tool options unions, all bind tree nodes, and the bind tree itself. These are needed for correctness checks during serialization tests.

Fixed input sequence stringification: a range constraint appearing after a group no longer fails to parse.

Fixed some outdated code files that broke after we removed `button_press_type` from modifier nodes.

Fixed mistakes in a few places where we accidentally serialized `size_t`s. That's a big no-no since that type has a platform-dependent size. (Sadly, we can't use templates to disallow serializing them, because `size_t` is just a typedef for `unsigned long long` in MSVC; we can't distinguish it from `uint64_t`.)

Fixed the default-comparable-anonymous-struct macros.

The bitstream classes now serialize QStrings as UTF-8 rather than as "whatever QString uses internally as of this program being compiled."

Added a missing function definition for `full_bitstream_position`, and marked some functions as nodiscard for better error checking.

---

With these changes, the program now builds again. Sadly, serialization tests fail, and because this is a bitstream, it's particularly challenging to figure out how. I have some redesign ideas that will make it harder to make mistakes, and easier to run correctness checks (including at compile-time). Besides, the current bitstream stuff has nothing in place for run-time error handling, which is a huge problem. It's not *terrible* for a semi-first draft but it's not great.

+++ native/src/DovahKit/editor/subsystems/worldinput2/input_sequence.inl
@@ -274,7 +274,7 @@ namespace dovahkit::subsystems::worldinput2 {
          stream.read(this->button.key.vk);
          stream.read(this->button.key.scan_code);
       } else {
-         size_t size;
+         uint32_t size;
          stream.read(size);
          this->children.resize(size);
          for (size_t i = 0; i < size; ++i) {
@@ -292,7 +292,7 @@ namespace dovahkit::subsystems::worldinput2 {
          stream.write(this->button.key.scan_code);
          assert(this->children.empty());
       } else {
-         size_t size = this->children.size();
+         uint32_t size = this->children.size();
          stream.write(size);
          if (size > 0) {
             for (const auto* g : this->children) {
